requirements.txt
altgraph==0.17.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
et_xmlfile==2.0.0
fire==0.7.1
fonttools==4.61.1
lxml==6.0.2
macholib==1.16.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pdf2docx==0.5.8
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==2.23
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
PyMuPDF==1.26.7
pypdf==6.6.0
pypdfium2==5.3.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6_sip==13.11.0
pytesseract==0.3.13
python-docx==1.2.0
python-pptx==1.0.2
reportlab==4.4.7
setuptools==80.9.0
termcolor==3.3.0
typing_extensions==4.15.0
xlsxwriter==3.2.9
